ner/bert_crf/requirements.txt
pymongo==3.10.1
stop-words==2018.7.23
pytorch-crf==0.7.2
blist==1.3.6
pip install https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.1/en_core_sci_md-0.5.1.tar.gz